// agent/src/common/generated/proto/key_pool.ts
export interface SignDigestRequest {
  /** / Default key: `SIGNING_KEY_SERVICE_RESPONSE` (used if left as `SIGNING_KEY_UNSPECIFIED`). */
  signingKey:
    | SigningKey
    | undefined;
  /**
   * / The actual digest bytes to sign. Must be exactly 32 bytes.
   * /
   * / When signing an Ethereum transaction or message,
   * / this should be the hash of the transaction or message, as per the Ethereum
   * / specification.
   */
  digest: Buffer;
}

export interface SignDigestResponse {
  signature: EcdsaSignature | undefined;
}

/** / Compute a signature on a message. */
export interface SignMessageRequest {
  /** / Default key: `SIGNING_KEY_SERVICE_RESPONSE` (used if left as `SIGNING_KEY_UNSPECIFIED`). */
  signingKey:
    | SigningKey
    | undefined;
  /** / Hash function to use to compute the digest to sign. */
  hashFunction: HashFunction;
  /** / The bytes of the message to sign. Maximum message size is 1Mib (2**20). */
  message: Buffer;
}

export interface SignMessageResponse {
  /**
   * / An signatuer using Ethereum's 65 byte format `r || s || v`, where `r` and `s`
   * / the signature components (each 32 bytes) and `v` is a zero or one byte
   * / used for key recovery.
   */
  signature: Buffer;
}

export interface SignEthereumTransactionRequest {
  /** / Default key: `SIGNING_KEY_ETHEREUM` (used if left as `SIGNING_KEY_UNSPECIFIED`). */
  signingKey:
    | SigningKey
    | undefined;
  /**
   * / RLP-encoded unsigned transaction.
   * / Must contain 6 or 9 element. If 9 elements are used, the 7th element
   * / it is assumed to be the chain ID and the transaction is signed using EIP-155;
   * / otherwise a legacy signature is used.
   */
  txData: Buffer;
}

export interface SignEthereumTransactionResponse {
  /**
   * / RLP-encoded signed transaction.
   * / Always 9 elements: the six elements from the input, followed by v, r, s.
   */
  txData: Buffer;
}

export interface GetEthereumAddressRequest {
  /** / Default key: `SIGNING_KEY_ETHEREUM` (used if left as `SIGNING_KEY_UNSPECIFIED`). */
  signingKey: SigningKey | undefined;
}

/** TODO: should use 0x prefix? should support ERC-55? */
export interface GetEthereumAddressResponse {
  /** / Hex encoded 40 bytes. */
  ethereumAddress: string;
}
// ...
export const SignMessageRequest: MessageFns<SignMessageRequest> = {
  encode(message: SignMessageRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signingKey !== undefined) {
      SigningKey.encode(message.signingKey, writer.uint32(10).fork()).join();
    }
    if (message.hashFunction !== 0) {
      writer.uint32(16).int32(message.hashFunction);
    }
    if (message.message.length !== 0) {
      writer.uint32(26).bytes(message.message);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SignMessageRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSignMessageRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signingKey = SigningKey.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.hashFunction = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.message = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SignMessageRequest {
    return {
      signingKey: isSet(object.signingKey) ? SigningKey.fromJSON(object.signingKey) : undefined,
      hashFunction: isSet(object.hashFunction) ? hashFunctionFromJSON(object.hashFunction) : 0,
      message: isSet(object.message) ? Buffer.from(bytesFromBase64(object.message)) : Buffer.alloc(0),
    };
  },

  toJSON(message: SignMessageRequest): unknown {
    const obj: any = {};
    if (message.signingKey !== undefined) {
      obj.signingKey = SigningKey.toJSON(message.signingKey);
    }
    if (message.hashFunction !== 0) {
      obj.hashFunction = hashFunctionToJSON(message.hashFunction);
    }
    if (message.message.length !== 0) {
      obj.message = base64FromBytes(message.message);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SignMessageRequest>, I>>(base?: I): SignMessageRequest {
    return SignMessageRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SignMessageRequest>, I>>(object: I): SignMessageRequest {
    const message = createBaseSignMessageRequest();
    message.signingKey = (object.signingKey !== undefined && object.signingKey !== null)
      ? SigningKey.fromPartial(object.signingKey)
      : undefined;
    message.hashFunction = object.hashFunction ?? 0;
    message.message = object.message ?? Buffer.alloc(0);
    return message;
  },
};

function createBaseSignMessageResponse(): SignMessageResponse {
  return { signature: Buffer.alloc(0) };
}

export const SignMessageResponse: MessageFns<SignMessageResponse> = {
  encode(message: SignMessageResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signature.length !== 0) {
      writer.uint32(10).bytes(message.signature);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SignMessageResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSignMessageResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signature = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SignMessageResponse {
    return { signature: isSet(object.signature) ? Buffer.from(bytesFromBase64(object.signature)) : Buffer.alloc(0) };
  },

  toJSON(message: SignMessageResponse): unknown {
    const obj: any = {};
    if (message.signature.length !== 0) {
      obj.signature = base64FromBytes(message.signature);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SignMessageResponse>, I>>(base?: I): SignMessageResponse {
    return SignMessageResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SignMessageResponse>, I>>(object: I): SignMessageResponse {
    const message = createBaseSignMessageResponse();
    message.signature = object.signature ?? Buffer.alloc(0);
    return message;
  },
};

function createBaseSignEthereumTransactionRequest(): SignEthereumTransactionRequest {
  return { signingKey: undefined, txData: Buffer.alloc(0) };
}

export const SignEthereumTransactionRequest: MessageFns<SignEthereumTransactionRequest> = {
  encode(message: SignEthereumTransactionRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signingKey !== undefined) {
      SigningKey.encode(message.signingKey, writer.uint32(10).fork()).join();
    }
    if (message.txData.length !== 0) {
      writer.uint32(18).bytes(message.txData);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SignEthereumTransactionRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSignEthereumTransactionRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signingKey = SigningKey.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.txData = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SignEthereumTransactionRequest {
    return {
      signingKey: isSet(object.signingKey) ? SigningKey.fromJSON(object.signingKey) : undefined,
      txData: isSet(object.txData) ? Buffer.from(bytesFromBase64(object.txData)) : Buffer.alloc(0),
    };
  },

  toJSON(message: SignEthereumTransactionRequest): unknown {
    const obj: any = {};
    if (message.signingKey !== undefined) {
      obj.signingKey = SigningKey.toJSON(message.signingKey);
    }
    if (message.txData.length !== 0) {
      obj.txData = base64FromBytes(message.txData);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SignEthereumTransactionRequest>, I>>(base?: I): SignEthereumTransactionRequest {
    return SignEthereumTransactionRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SignEthereumTransactionRequest>, I>>(
    object: I,
  ): SignEthereumTransactionRequest {
    const message = createBaseSignEthereumTransactionRequest();
    message.signingKey = (object.signingKey !== undefined && object.signingKey !== null)
      ? SigningKey.fromPartial(object.signingKey)
      : undefined;
    message.txData = object.txData ?? Buffer.alloc(0);
    return message;
  },
};

function createBaseSignEthereumTransactionResponse(): SignEthereumTransactionResponse {
  return { txData: Buffer.alloc(0) };
}

export const SignEthereumTransactionResponse: MessageFns<SignEthereumTransactionResponse> = {
  encode(message: SignEthereumTransactionResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.txData.length !== 0) {
      writer.uint32(10).bytes(message.txData);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SignEthereumTransactionResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSignEthereumTransactionResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.txData = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SignEthereumTransactionResponse {
    return { txData: isSet(object.txData) ? Buffer.from(bytesFromBase64(object.txData)) : Buffer.alloc(0) };
  },

  toJSON(message: SignEthereumTransactionResponse): unknown {
    const obj: any = {};
    if (message.txData.length !== 0) {
      obj.txData = base64FromBytes(message.txData);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SignEthereumTransactionResponse>, I>>(base?: I): SignEthereumTransactionResponse {
    return SignEthereumTransactionResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SignEthereumTransactionResponse>, I>>(
    object: I,
  ): SignEthereumTransactionResponse {
    const message = createBaseSignEthereumTransactionResponse();
    message.txData = object.txData ?? Buffer.alloc(0);
    return message;
  },
};
// ...
export type KeyPoolServiceService = typeof KeyPoolServiceService;
// ...
export interface KeyPoolServiceServer extends UntypedServiceImplementation {
  signDigest: handleUnaryCall<SignDigestRequest, SignDigestResponse>;
  signMessage: handleUnaryCall<SignMessageRequest, SignMessageResponse>;
  signEthereumTransaction: handleUnaryCall<SignEthereumTransactionRequest, SignEthereumTransactionResponse>;
  getEthereumAddress: handleUnaryCall<GetEthereumAddressRequest, GetEthereumAddressResponse>;
}
// ...
type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };